use complete pad spec in string tests
The :pad :fill field was not set in most string tests.
When set it was set as a string.

The current format-string function accepts this,
however, :fill is meant to be a required, int32 field.

This commit corrects the usage of fill,
so tests won't fail due to missing :fill on future implementations
of format-string.