add helper macro for padding generators
This commit adds the `pad-around` macro which reads a pad spec
then yields characters around an expression accordingly.
Abstracting this behavior helps reduce repetitive code and
keep behavior consistent between formatters.
It may also be a useful utility for custom formatters.

This commit fixes a bug in format-integer where content was biased right
when aligned center, which was incosistent with other formatters.